4.1
客户登记

+++ xinnongreniOS/Classes/Home/View/XNRSpecialCell.m
@@ -161,7 +161,7 @@
         self.shopcarBtn.hidden = YES;
     }else{
         self.priceLabel.textColor = R_G_B_16(0xff4e00);
-        self.priceLabel.text = [NSString stringWithFormat:@"%.2f",model.unitPrice.floatValue];
+        self.priceLabel.text = [NSString stringWithFormat:@"%.2f",model.unitPrice.doubleValue];
         self.shopcarBtn.hidden = NO;
     }
 }
